Removed redundant files and created new HTML and CSS files for website
The commit includes creating new HTML files for the cyber-security website. The HTML pages are 'ClientLogin', '404', 'About', 'index', and 'Contact'. Alongside this, the new main CSS file was also added, which includes styling related to fonts, colors, layout, and responsive design to make the website visually appealing. The 404 error page was also added to enhance user experience. The changes aim to establish a more interactive and effective web presence for the cyber-security startup.